Added: PlayerView.Translate() now checks if direction was changed and rotates accordingly Moved: Hitbox size and center, hitbox gizmos moved to PlayerView
